Fix a bug where `pebble install --logs` would not show logs
To support lightblue we introduced a pebble.install_app() method which dispatches
to a bluetooth installer or a websocket installer. We need to return the success value
of the underlying implementation to the caller.

I have added a FIXME for the bluetooth implementation. Instead of returning False it
raises Exception. We will have to decide on one way or the other and use the same
error mechanism for both implementation.